Code aufgeräumt (Benamsung etc)
